refactor non website content
