+++ src/components/faq/Cmdf2021.jsx
@@ -5,7 +5,7 @@ import {
   Columns,
   Column,
 } from '@lib/Containers'
-import ExpandableFaqList from '@components/faqTemplates/shared/FaqList'
+import ExpandableFaqList from '@components/faq/shared/FaqList'
 import backgroundPlants from '@assets/faq__background_plants.svg'
 import titleOptions from '@assets/faq__title_options.svg'
 import questionLeftDecor from '@assets/faq__question_leftdecor.svg'
@@ -110,7 +110,7 @@ const SectionContainerWithBackground = styled(SectionContainerWithContainBackgro
 // FAQ Section with two columns and scattered layout:
 // General    Logistics
 //          Teams & Projects
-const ExpandableScatteredCategories = ({ config }) => {
+const ExpandableScatteredCategories = ({ config, shouldDisplay }) => {
   const [categorizedFaqMap, setFaqMap] = useState(new Map())
   const [faq, setFaq] = useState([])
 
@@ -133,7 +133,7 @@ const ExpandableScatteredCategories = ({ config }) => {
     categorizeFaq(faq)
   }, [faq])
 
-  return (
+  return shouldDisplay ? (
     <SectionContainerWithBackground
       width={config.containerWidth}
       margin={config.containerMargin}
@@ -183,7 +183,7 @@ const ExpandableScatteredCategories = ({ config }) => {
         </Column>
       </Columns>
     </SectionContainerWithBackground>
-  )
+  ) : null
 }
 
 export default ExpandableScatteredCategories
