+++ src/components/faq/ExpandableWithCategories.jsx
