+++ src/components/faq/shared/ExpandableFaq.jsx
